Avoid converting mount path specified by user
The parse_cmdline() function was revised to avoid converting the
mount path specified by the user into a canonicalized absolute
path. This allows the user to specify a relative mount path to
avoid the race conditions that may be associated with the use of an
equivalent absolute path.

Commit UUID: e92ab761-fc24-4d92-b001-09432b37ac2a